Hide/show automatically pass list and pass view
Automatically hide the pass list and the pass view when there is no pass to
show. To do that, the main window (leaflet) will be forced to fold and the
empty pass list will be hidden under a status page.

As soon as the pass list has passes again, the window (leaflet) will be allowed
to unfold and the pass list widget will be put on top of the status page.