Improve barcode drawing
Barcodes will be drawn as large as possible (as long as they fit in the
available space).

The drawing function is now taking into account the size of the margins
to make sure that there is always enough space around the barcode.

Performed other small corrections to the drawing algorithm.