model: install private library in private folder
`libbarcode-content-encoder.so` is actually a private library, only
usable by `passes` (no development headers, no ABI version tracking). As
such, it should be installed in a private folder instead of the system's
`libdir`.

This change also requires using the full path to the library in
`barcode_content_encoder.py` as the aforementioned private folder isn't
part of the standard library search path.